Refactor types, separate helpers
