+++ src/components/breadcrumb/Breadcrumb.vue
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import { Breadcrumb as ShadcnBreadcrumb, BreadcrumbList as ShadcnBreadcrumbList } from '@/shadcn/ui/breadcrumb';
 import BreadcrumbItem from '@/components/breadcrumb/BreadcrumbItem.vue';
-import type { BreadcrumbItem as BreadcrumbItemType } from '@/types/components.ts';
+import type { BreadcrumbItem as BreadcrumbItemType } from '@/types';
 
 const props = defineProps<{
   items: BreadcrumbItemType[];
