+++ src/helpers/routes.ts
@@ -1,0 +1,9 @@
+import type { RouteRecord } from '@/types';
+
+export function createRoute(route: RouteRecord): RouteRecord {
+  return {
+    showInMenu: true, // Default value
+    menuLabel: route.name?.toString() || '', // Default label from name
+    ...route, // This allows overriding the default
+  };
+}
