+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export type { BreadcrumbItem } from './breadcrumb';
+export type { RouteRecord } from './routes';
