+++ src/types/routes.ts
@@ -5,10 +5,4 @@ export type RouteRecord = RouteRecordRaw & {
   menuLabel?: string;
 };
 
-export function createRoute(route: RouteRecord): RouteRecord {
-  return {
-    showInMenu: true, // Default value
-    menuLabel: route.name?.toString() || '', // Default label from name
-    ...route, // This allows overriding the default
-  };
-}
+
